Fix color-related terminal todo!
